# src/lib/sdk/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Scalius Commerce API",
    "version": "1.0.0",
    "description": "API for the Scalius Commerce platform. This specification provides all the necessary details to build a complete e-commerce storefront, from displaying products and categories to processing orders and managing customer data.",
    "contact": {
      "name": "Scalius Support",
      "url": "https://scalius.com/support"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A short-lived JSON Web Token obtained from the /auth/token endpoint."
      },
      "apiToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token",
        "description": "A static, pre-shared token for machine-to-machine authentication to obtain a JWT."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error Type"
          },
          "message": {
            "type": "string",
            "example": "A detailed error message."
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/locations/cities": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "Get all delivery cities",
        "description": "Returns a list of all active cities available for delivery. This is the first step in a hierarchical location selector (City -> Zone -> Area).",
        "responses": {
          "200": {
            "description": "A list of active cities.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "example": "city"
                          },
                          "parentId": {
                            "type": "string",
                            "nullable": true
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "sortOrder": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/locations/zones": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "Get delivery zones for a city",
        "description": "Returns a list of all active zones for a specified city ID. Call this after the user has selected a city.",
        "parameters": [
          {
            "name": "cityId",
            "in": "query",
            "required": true,
            "description": "The ID of the parent city.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of active zones for the specified city.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "example": "zone"
                          },
                          "parentId": {
                            "type": "string"
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "sortOrder": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/locations/areas": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "Get delivery areas for a zone",
        "description": "Returns a list of all active areas for a specified zone ID. Call this after the user has selected a zone. This step may be optional depending on `fieldVisibility` settings.",
        "parameters": [
          {
            "name": "zoneId",
            "in": "query",
            "required": true,
            "description": "The ID of the parent zone.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of active areas for the specified zone.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "example": "area"
                          },
                          "parentId": {
                            "type": "string"
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "sortOrder": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
...
  }
}